Move mailbox status to mailbox class
This allows to be cleaner, and more importantly to store the mailbox_status in an instance variable that can directly be reused later.

This is ok because almost each operation spawns a new HeliotropeFakeMailbox, so the status would be valid between this moment and the moment the operation is done. However, if an operation appends in between, we're screwed.